refactor(review): extract CheckRunManager from ReviewOrchestrator (#250)
Step 2 of the decomposition: move the check-run create/update subsystem
into its own @ApplicationScoped collaborator. Pure extract-class, no
behaviour change — createCheckRun, updateCheckRun (with the
completion-only retry after a failed full update), githubTimestamp, and
the CheckRunUpdate record move verbatim.

The orchestrator injects CheckRunManager and delegates its three call
sites (in-progress create, completion update, failure-path update). With
the CI subsystem already gone (step 1), checkRunClient is no longer used
by the orchestrator, so it (and the time + CHECK_STATUS_IN_PROGRESS
constants) leave too.

Tests: the CheckRunUpdates cases move to a new CheckRunManagerTest (3
cases, exercising the manager directly with the mocked client);
ReviewOrchestratorTest builds a real CheckRunManager from the mocked
client so the review() integration path stays covered. Full suite green,
spotbugs + spotless clean.
